Improve siva index generation in ReadWriter
When using a ReadWriter the index in regenerated each time Index is
called. These are the steps to generate one usable index:

* Create new index merging index from file and current changes
* Remove duplicates
* Sort index by position

Also, to find a file in the index it is walked until a match is found.
This needs to be done each time a file has to be opened.

For small number of files this is OK but when a repo has a lot or
references the time spent here can be a lot.

Now there's a new index type called OrderedIndex that stores the
IndexEntries in lexicographic order. This allows to do binary searches
for faster file location and also makes possible update the index
instead of regenerating it each time.

Signed-off-by: Javi Fontan <jfontan@gmail.com>